removed dependency on jackson and kept only gson
this comes at cost of not using final fields for DTOs but in absence of
setters I think that is acceptable
Alternatively custom deserialization can be implemented but it seems
like that is not worth the effort and extra code
there is small performance degradation in calls to get http protocol.

All of the above can be fixed but I do not think it is worth the extra
code.